S3 README.md
Create Bucket: The create_bucket function creates an S3 bucket with the specified name.

Upload File: The upload_file function uploads a local file to the specified S3 bucket. If object_name is not provided, the file will be uploaded with the same name as the local file.

List Files: The list_files function lists all the files in the specified S3 bucket.

Download File: The download_file function downloads a file from the specified S3 bucket. If file_name is not provided, the file will be downloaded with the same name as it has in the bucket